Add files via upload
Remember this is just a tutorial that I did. This is meant to be more of a reference to lead us in the right direction for how to update dives, edit dives, delete dives or add a new dive.  The tutorial that I did was based off of a note taking app.