Add files via upload
DivingInformation1 has been updated with code and XML, OtherDivingDetails has also been updated. I am having problems with some of the buttons. When I go to wildlife, equipment or divebuddy and I press back on either of the three the application crashes. It was not doing that before and I checked several times to make sure the navigation is right and everything seems to be right. I think it has something to do with "ID" when I checked the terminal in the "RUN" section. I tried to do some of the logic but I could not get to it the way I wanted. For the files you can always just create them and then copy and paste the code or whatever works.